+++ .github/workflows/test.yaml
@@ -75,8 +75,8 @@ jobs:
       TWILIO_AUTH_TOKEN: 123
       elasticURL: "http://localhost:9200"
       NODE_COVERALLS_DEBUG: 1
-      NODE_ENV: dev
-      DEV: true
+      NODE_ENV: TEST
+      TEST: true
     steps:
       - uses: actions/checkout@v2
       - name: install node v14.19
@@ -96,7 +96,7 @@ jobs:
       # Add two sections - one which shouldn't be removed, one which should
       - run: psql -U postgres -h localhost -p 5432 -d searchneu_dev -c "INSERT INTO sections (class_hash, id, crn) VALUES ('neu.edu/202240/CS/2501', 'neu.edu/202240/CS/2501/123456789', '123456789');"
       - run: psql -U postgres -h localhost -p 5432 -d searchneu_dev -c "INSERT INTO sections (class_hash, id, crn, last_update_time) VALUES ('neu.edu/202240/CS/2501', 'neu.edu/202240/CS/2501/987654321', '987654321', '1999-04-03 18:34:35.882');"
-      - run: UPDATE_ONLY_ONCE=true LOG_LEVEL=VERBOSE yarn updater
+      - run: LOG_LEVEL=VERBOSE yarn updater
 
       - name: Checks that 'yarn dev' actually starts the API server
         run: curl -Is http://localhost:4000/.well-known/apollo/server-health | head -1 | grep -q 200
